Fixed small bugs

+++ Dtos/PeeringTask/GetPeeringTaskStudentGradesDtoResponse.cs
@@ -8,7 +8,7 @@ namespace patools.Dtos.Task
 {
     public class GetPeeringTaskStudentGradesDtoResponse
     {
-        public int? MinGrade { get; set; }
+        public float? MinGrade { get; set; }
         public float? MaxGrade { get; set; }
         public List<GetPeeringTaskCoordinatesDtoResponse> Coordinates { get; set; }
     }
